+++ src/pages/ContactUsPage.jsx
@@ -3,7 +3,6 @@ import React, { useState } from "react";
 import axios from "axios";
 import { API_BASE_URL } from "../constants/apiConstants";
 
-import ContactUsAlert from '../components/contactUsAlert';
 
 import NavigationHeader from "../components/navigationHeader.jsx";
 import MarbleBackground from "../img/MarbleBackground.png";
@@ -15,6 +14,13 @@ import Button from "@material-ui/core/Button";
 import MapContainer from "../components/contactUs/googleMap";
 import FormControl from "@material-ui/core/FormControl";
 import TextareaAutosize from "@material-ui/core/TextareaAutosize";
+
+
+
+import Dialog from '@material-ui/core/Dialog';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import MessageSentAnimation from '../img/messageSentAnimation.gif';
 
 import { makeStyles } from "@material-ui/core/styles";
 
@@ -70,6 +76,22 @@ const useStyles = makeStyles((theme) => ({
     margin: "50px auto 200px auto"
   },
 
+  title: {
+    width: "511px",//72px
+    height: "511px",
+    textAlign: "center",
+    fontFamily: "Playfair Display",
+    fontStyle: "normal",
+    fontWeight: "bold",
+    fontSize: "100px",
+    
+    background: "#F5E9DE",
+},
+
+conText:{
+  color: "#BF7950",
+},
+
 }));
 
 const ContactUsPage = () => {
@@ -94,14 +116,22 @@ const ContactUsPage = () => {
     axios
       .post( API_BASE_URL + "/sendMail", email )
       .then((response) => {
-        //console.log(response);        
+        console.log(response);        
       })
       .catch((err) => {
-        //console.log(err);
+        console.log(err);
     });
   }
   
+  const [open, setOpen] = React.useState(false);
 
+    const handleClickOpen = () => {
+        setOpen(true);
+    };
+
+    const handleClose = () => {
+        setOpen(false);
+    };
   return (
     <div className={classes.backgroundImage}>
       <NavigationHeader />
@@ -115,10 +145,47 @@ const ContactUsPage = () => {
 
           <Grid item xs={12}>
             <div className={classes.form}>
+              <div>
+              <Dialog
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+        //className={classes.alertBox}
+      >
+        <Grid container className={classes.title} >
+
+        <Grid item xs={12}>
+          <DialogContent>
+            <DialogContentText id="alert-dialog-description">
+              <Typography variant="h5" className={classes.conText}>Your message is sent</Typography>
+            </DialogContentText>
+          </DialogContent>
+        </Grid>
+
+        <Grid item xs={12}>
+          <DialogContent>
+            <DialogContentText id="alert-dialog-description">
+              <img src={MessageSentAnimation} alt="Message Sent Animation Gif"/>
+            </DialogContentText>
+          </DialogContent>
+        </Grid>
+
+        <Grid item xs={12}>
+          <DialogContent>
+            <DialogContentText id="alert-dialog-description" display="block">
+              We appreciate your feedback. We will get intouch with you as soon as possible.
+            </DialogContentText>
+          </DialogContent>
+        </Grid>
+
+        </Grid>
+      </Dialog>
+              </div>
               <form className={classes.root} noValidate autoComplete="off" onSubmit={handleSubmit}>
                 <FormControl>
                   <Grid container>
-                    <ContactUsAlert />
+                    
                     <Grid item xs={12}>
                       <div className={classes.nameInput}>
                         <TextField
@@ -185,6 +252,7 @@ const ContactUsPage = () => {
                         color="primary"
                         className={classes.submit}
                         onSubmit={handleSubmit}
+                        onClick={handleClickOpen}
                       >
                         Send Message
                       </Button>
